[updates-e2e] run tests from a single root file
